Set of functions defining special simulation methods for column and its type

Whenever there's a need to simulate column using specific function (as a spec parameter in YAML configuration file), such method should be defined in one of opt_simul_spec_<column_type> functions.

opt_simul_spec_character(name = simul_spec_character_name, ...)

opt_simul_spec_numeric(distr = simul_spec_numeric_distr, ...)

opt_simul_spec_integer(distr = simul_spec_integer_distr, ...)

opt_simul_spec_logical(distr = simul_spec_logical_distr, ...)

opt_simul_spec_date(distr = simul_spec_date_distr, ...)

Arguments

name

Function for simulating personal names.

...

Other custom special methods.

distr

Function for simulating data from desired distribution.

Details

Currently defined special methods are:

name For character column, that allows to simulate character reflecting real names and surnames

distr For all the remaining column types. The method allows to simulate data with specified distribution generator, such as rnorm, rbinom etc.

Each 'spec' method receives n parameter (the desired number of rows to simulate), all the default column-based parameters (type, unique, not_null, etc.) but also a special one named spec_params that are applied to selected distribution simulation method.

See for example simul_spec_character_name definition.